Split widget: hot badge
